+++ app/(app)/contribute/page.tsx
@@ -1,0 +1,5 @@
+export default function ContributePage() {
+  return (
+    <div className='flex h-full items-center justify-center'>Contribute Page</div>
+  );
+}