+++ app/layout.tsx
@@ -1,15 +1,9 @@
 import type { Metadata } from 'next';
-import { Geist, Rethink_Sans } from 'next/font/google';
+import { Geist } from 'next/font/google';
 import '@/styles/globals.css';
-import Navbar from '@/components/Navbar';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
-  subsets: ['latin']
-});
-
-const rethinkSans = Rethink_Sans({
-  variable: '--font-rethink-sans',
   subsets: ['latin']
 });
 
@@ -25,12 +19,8 @@ export default function RootLayout({
 }) {
   return (
     <html lang='en'>
-      <body
-        className={`${geistSans.variable} ${rethinkSans.variable} min-h-screen bg-white text-gray-900 antialiased`}
-      >
-        <Navbar />
-
-        <main className='h-[calc(100vh-4rem)] w-full'>{children}</main>
+      <body className={`${geistSans.variable} light min-h-screen antialiased`}>
+        <main>{children}</main>
       </body>
     </html>
   );